Rename botchat-core to botchat-minimal
