+++ packages/bundle/src/index-minimal.ts
