+++ samples/minimal-bundle-with-markdown/index.html
@@ -10,7 +10,7 @@
 
       And then, we bring Markdown-It back to the project.
     -->
-    <script src="https://cdn.botframework.com/botframework-webchat/preview/botchat-core.js"></script>
+    <script src="https://cdn.botframework.com/botframework-webchat/preview/botchat-minimal.js"></script>
     <script src="https://unpkg.com/markdown-it@8.4.2/dist/markdown-it.min.js"></script>
     <style>
       html, body { height: 100% }
