+++ samples/minimal-bundle/index.html
@@ -11,7 +11,7 @@
       However, it contains:
       - Direct Line JS
     -->
-    <script src="https://cdn.botframework.com/botframework-webchat/preview/botchat-core.js"></script>
+    <script src="https://cdn.botframework.com/botframework-webchat/preview/botchat-minimal.js"></script>
     <style>
       html, body { height: 100% }
       body { margin: 0 }
