+++ samples/speech-browser/serve.json
@@ -2,6 +2,6 @@
   "public": "../../",
   "rewrites": [
     { "source": "/", "destination": "samples/speech-browser/index.html" },
-    { "source": "/botchat-core.js", "destination": "packages/bundle/dist/botchat-core.js" }
+    { "source": "/botchat-minimal.js", "destination": "packages/bundle/dist/botchat-minimal.js" }
   ]
 }
