+++ serve.json
@@ -2,7 +2,7 @@
   "redirects": [
     { "source": "/", "destination": "samples/full-bundle/index.html" },
     { "source": "/botchat.js", "destination": "packages/bundle/dist/botchat.js" },
-    { "source": "/botchat-core.js", "destination": "packages/bundle/dist/botchat-core.js" },
-    { "source": "/botchat-es5.js", "destination": "packages/bundle/dist/botchat-es5.js" }
+    { "source": "/botchat-es5.js", "destination": "packages/bundle/dist/botchat-es5.js" },
+    { "source": "/botchat-minimal.js", "destination": "packages/bundle/dist/botchat-minimal.js" }
   ]
 }
